Fix merge of 3.7 to master
